Move all the version-related code into the bag_versioner
This reduces the size of the common lib, and means we'll have less
CI churn if we're changing the versioning code (as I'm about to do!).
Changes to the common lib trigger a complete rebuild of the whole
storage service, which isn't necessary for changing this bit of the code.
