+++ bag_versioner/docker-compose.yml
@@ -1,5 +1,9 @@
 sqs:
-  image: s12v/elasticmq
+  image: "s12v/elasticmq"
   ports:
     - "9324:9324"
     - "4789:9324"
+dynamodb:
+  image: "peopleperhour/dynamodb"
+  ports:
+    - "45678:8000"
