+++ bag_versioner/src/main/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/IngestVersionManager.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.versioning
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning
 
 import java.time.Instant
