+++ bag_versioner/src/main/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/VersionPickerError.scala
@@ -1,7 +1,5 @@
 package uk.ac.wellcome.platform.storage.bag_versioner.versioning
 
-import uk.ac.wellcome.platform.archive.common.versioning.IngestVersionManagerError
-
 sealed trait VersionPickerError
 
 case class InternalVersionPickerError(e: Throwable) extends VersionPickerError
