+++ bag_versioner/src/main/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/dynamo/DynamoIngestVersionManager.scala
@@ -1,0 +1,6 @@
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning.dynamo
+
+import uk.ac.wellcome.platform.storage.bag_versioner.versioning.IngestVersionManager
+
+class DynamoIngestVersionManager(val dao: DynamoIngestVersionManagerDao)
+    extends IngestVersionManager
