+++ bag_versioner/src/main/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/dynamo/DynamoVersionRecord.scala
@@ -1,10 +1,11 @@
-package uk.ac.wellcome.platform.archive.common.versioning.dynamo
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning.dynamo
 
 import java.time.Instant
 
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagVersion
 import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
-import uk.ac.wellcome.platform.archive.common.versioning.VersionRecord
+import uk.ac.wellcome.platform.archive.common.storage.models.dynamo.DynamoID
+import uk.ac.wellcome.platform.storage.bag_versioner.versioning.VersionRecord
 import uk.ac.wellcome.storage.dynamo.DynamoHashRangeKeyPair
 
 import scala.util.Try
