+++ bag_versioner/src/main/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/memory/MemoryIngestVersionManager.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning.memory
+
+import uk.ac.wellcome.platform.storage.bag_versioner.versioning.IngestVersionManager
+
+class MemoryIngestVersionManager(
+  val dao: MemoryIngestVersionManagerDao = new MemoryIngestVersionManagerDao()
+) extends IngestVersionManager
