+++ bag_versioner/src/main/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/memory/MemoryIngestVersionManagerDao.scala
@@ -1,9 +1,9 @@
-package uk.ac.wellcome.platform.archive.common.versioning.memory
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning.memory
 
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
 import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.platform.archive.common.storage.models.StorageSpace
-import uk.ac.wellcome.platform.archive.common.versioning.{
+import uk.ac.wellcome.platform.storage.bag_versioner.versioning.{
   IngestVersionManagerDao,
   VersionRecord
 }
