+++ bag_versioner/src/test/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/dynamo/DynamoIngestVersionManagerTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.versioning.dynamo
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning.dynamo
 
 import org.scanamo.auto._
 import org.scanamo.time.JavaTimeFormats._
@@ -6,7 +6,7 @@ import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
 import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.platform.archive.common.storage.models.StorageSpace
-import uk.ac.wellcome.platform.archive.common.versioning.{
+import uk.ac.wellcome.platform.storage.bag_versioner.versioning.{
   IngestVersionManager,
   IngestVersionManagerTestCases,
   VersionRecord
