+++ bag_versioner/src/test/scala/uk/ac/wellcome/platform/storage/bag_versioner/versioning/memory/MemoryIngestVersionManagerDaoTest.scala
@@ -1,7 +1,7 @@
-package uk.ac.wellcome.platform.archive.common.versioning.memory
+package uk.ac.wellcome.platform.storage.bag_versioner.versioning.memory
 
 import uk.ac.wellcome.fixtures.TestWith
-import uk.ac.wellcome.platform.archive.common.versioning.{
+import uk.ac.wellcome.platform.storage.bag_versioner.versioning.{
   IngestVersionManagerDao,
   IngestVersionManagerDaoTestCases,
   VersionRecord
