--- common/src/main/scala/uk/ac/wellcome/platform/archive/common/versioning/memory/MemoryIngestVersionManager.scala
@@ -1,7 +1,0 @@
-package uk.ac.wellcome.platform.archive.common.versioning.memory
-
-import uk.ac.wellcome.platform.archive.common.versioning.IngestVersionManager
-
-class MemoryIngestVersionManager(
-  val dao: MemoryIngestVersionManagerDao = new MemoryIngestVersionManagerDao()
-) extends IngestVersionManager
